A WIP commit of changes to enable the mesh transformations for items and ik-joints.
Notable changes: The new format uses c++ components which should most
likely be released in a library for qml. They use also some private headers
so check out the pro-example.pro and main-example.cpp in resources/

Known issues: The released qml-file has image folder in the beginning,
that doesn't work in desktop as it should. IIRC it worked in Android and
has been bugging me since.

In Spine dragon animation example I think there was some bug lingering in
how certain angles are animated, so all goes as it should until it decides
that the last part of the rotation must be done in wrong direction.

Ik-joint has some algorithm bug. Probably not a big deal to fix.